Consolidate torchrun acknowledgement values
